% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLandClim.r
\docType{data}
\name{default_config}
\alias{default_config}
\title{Create theoretical landscapes to be used in LandClim}
\format{
An object of class \code{data.frame} with 1 rows and 139 columns.
}
\usage{
default_config
}
\arguments{
\item{overwrite}{defult is FALSE, indicating that if the xml files already exist they will not been overwritten. A warning appears as a function return.}

\item{config}{the path to the configuration file as a character or a data.frame where each of the variables/columns is a tag in the model configuration file in LandClim. If not provided the default values will be used.}

\item{barkbeetle}{the path to the barkbeetle file as a character or a data.frame where each of the variables/columns is a tag in the bark beetle configuration file in LandClim. If not provided the default values will be used. The created file will be saved in the indicated workspacepath.}

\item{landtypeparameters}{the path to the landtypeparameters file as a character or a data.frame where each of the variables/columns is a tag in the land type parameters configuration file in LandClim. If not provided the default values will be used. The created file will be saved in the indicated workspacepath.}

\item{plantingparameters}{the path to the plantingparametersfile as a character or a data.frame where each of the variables/columns is a tag in the planting parameters configuration file in LandClim. If not provided the default values will be used. The created file will be saved in the indicated workspacepath.}

\item{randomstate}{the path to the randomstate file as a character or a data.frame where each of the variables/columns is a tag in the random state configuration file in LandClim. If not provided the default values will be used. The created file will be saved in the indicated workspacepath.}

\item{species}{the path to the species file as a character or a data.frame where each of the variables/columns is a tag in the species parameters configuration file in LandClim. If not provided the default values will be used. The created file will be saved in the indicated workspacepath.}

\item{workspacePath}{path where the the input xml files and output will be created. The name of the files will be the default names (config.xml, barkbeetle.xml, landtype.xml, planting.xml, randomstate.xml, species.xml) unless they have previously created.}

\item{consoleOutputPath}{path to the folder and name ( e.g. nameofthefile.txt) of the file where you want to storage the console outputs shown during the simulation. If not provided the console outcomes will be seen in the console and there will not be storaged.}

\item{binPath}{path to the folder where LandClim model binary files needed for your operative system are storaged.}
}
\value{
Run LandClim with the desired input files and saved in the \code{workspacePath} provided.
}
\description{
Create theoretical landscapes to be used in LandClim
}
\examples{
runLandClim(overwrite = FALSE,
            config = NA,
            barkbeetle = NA,
            landtypeparameters = NA,
            plantingparameters = NA,
            randomstate =  NA,
            species = NA,
            workspacePath = "workspace",
            consoleOutputPath = "",
            binPath = "bin")
}
\keyword{datasets}
