% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLandscape.R
\name{writeLandscape}
\alias{writeLandscape}
\title{Create theoretical landscapes to be used in LandClim}
\usage{
writeLandscape(
  ySize,
  xSize,
  cellSize,
  coor,
  altitude,
  slop,
  so,
  landt,
  asp,
  plot,
  fullpath
)
}
\arguments{
\item{ySize}{y landscape length in m.}

\item{xSize}{x landscape length in m.}

\item{cellSize}{cell size in m.}

\item{coor}{CRS, coordinate system in EPSG.}

\item{altitude}{Altitude in m, if it is a single value the whole map will have the same value, if you want to provide different values per cell you have to provide a vector of length equal to the number of cells and consider that the values will be assigned from top to bottom and from left to right.}

\item{slop}{Slope in degrees, if it is a single value the whole map will have the same value, if you want to provide different values per cell you have to provide a vector of length equal to the number of cells and consider that the values will be assigned from top to bottom and from left to right.}

\item{so}{bucket size of the soil, if it is a single value the whole map will have the same value, if you want to provide different values per cell you have to provide a vector of length equal to the number of cells and consider that the values will be assigned from top to bottom and from left to right.}

\item{landt}{Land type,if it is a single value the whole map will have the same value, if you want to provide different values per cell you have to provide a vector of length equal to the number of cells and consider that the values will be assigned from top to bottom and from left to right.}

\item{asp}{aspect in degrees, if it is a single value the whole map will have the same value, if you want to provide different values per cell you have to provide a vector of length equal to the number of cells and consider that the values will be assigned from top to bottom and from left to right.}

\item{plot}{If TRUE the created landscapes will be plotted as an extra outcome.}

\item{fullpath}{The full path where the landscapes will be saved.}
}
\value{
Save the created maps into the \code{fullpath} provided.
}
\description{
Create theoretical landscapes to be used in LandClim
}
\examples{
theorLandscape (ySize = 1500, xSize = 1500, cellSize = 30, coor = 5514,
                altitude = 1800, slop = 0, so = 20, landt = 11, asp = 0,
                plot = TRUE, fullpath = "/Users/olalladiaz/MyRepos/beetleunc/Data/theoretical/squared"=
}
